V4 weight function added
It now contains the function of weight with the units: kgs, tons, lbs, and oz. 
It has a different layout in the coding as the first function, and it's mostly as a learning experience. So it definitely was a learning experience to know that I wrote so much code in the first function and it was not needed.